% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{baseline}
\alias{baseline}
\title{Baseline Correction.}
\usage{
baseline(df, baseline_length = 4, met_data, tz = "UTC")
}
\arguments{
\item{df}{Data frame containing velocity data as produced by solve_hrm()}

\item{baseline_length}{Optional minimum length for a baseline period to be considered (default 4 hours)}

\item{met_data}{Data frame containing met data at half hourly intervals (e.g., 10:00:00, 10:30:00, 11:00:00).
Required columns are Timestamp (POSIXct format: YYYY-MM-DD HH:MM:SS, in either UTC or local time), Tair (ºC),
PAR (W m-2 or µmol m-2 s-1), RH (\%), and optionally VPDair (kPa)}

\item{tz}{local timezone for the data}
}
\value{
Data frame containing original and corrected velocity data, matched with met data
}
\description{
Aligns sap velocity data with met data. Locates periods of 0 VPD, PAR, and
rainfall and uses those periods to assume zero flow. Applies correction to sap velocity data to account for
misalignment of probes/low flow drift.
}
\author{
Kenneth Davidson
}
