% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_temp}
\alias{convert_temp}
\title{Function to converts from resistance to temp in deg C}
\usage{
convert_temp(x)
}
\value{
Temp_C the measurement temperature in degC
}
\description{
Function to converts from resistance to temp in deg C
}
\author{
Kenneth Davidson
}
