% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_nearest}
\alias{find_nearest}
\title{A helper function for baseline}
\usage{
find_nearest(dat)
}
\description{
A helper function which finds the nearest non-NA value and applies it
}
\author{
https://stackoverflow.com/a/10081991
}
