% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getmode}
\alias{getmode}
\title{A helper function to find mode value to establish a pre-heat
baseline temp for QAQC and data processing}
\usage{
getmode(v)
}
\description{
A helper function to find mode value to establish a pre-heat
baseline temp for QAQC and data processi
}
\author{
Kenneth Davidson
}
