% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_dissapation_mode.R
\name{heat_dis_mode}
\alias{heat_dis_mode}
\title{This function determines the time it takes in seconds from max heat pulse
back to the appropriate baseline temperature}
\usage{
heat_dis_mode(df, Dates = unique(df$Date), Time_Post_Heat = c(30:180))
}
\arguments{
\item{df}{The input data frame}

\item{Dates}{The dates / date range to process}

\item{Time_Post_Heat}{The time in seconds it takes for the sap temperature to
return to the baseline temperature}
}
\value{
Heat_Dis_Data The processed data
}
\description{
Finds the baseline temp before the heat pulse, the average
increase in temperature from the baseline for a given period to time
(30-180 sec post heat by default), the time of peak heating, and the time
at which the heat returns to baseline
}
\author{
Kenneth Davidson
}
