% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Function to load and format raw CR1000 data}
\usage{
load_data(
  filepath,
  Origin_DateTime = "2000-01-01 01:00:00",
  End_DateTime = "2030-01-01 01:00:00"
)
}
\arguments{
\item{filepath}{The file path}

\item{Origin_DateTime}{e.g. 2000-01-01 01:00:00}

\item{End_DateTime}{e.g. 2030-01-01 01:00:00}
}
\value{
A dataframe in standard NGSapFlux_r format.
These data are not yet processed or QAQC'ed
}
\description{
function to load raw CR1000 data.
Set up to run for each CR1000 box's data in a separate directory.
Can be changed to process all boxes from one directory preferred
}
\author{
Kenneth Davidson

Jeremiah Anderson

Shawn Serbin
}
