% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{qaqc_remove}
\alias{qaqc_remove}
\title{Replaces values that were flagged in the fine QAQC with NA}
\usage{
qaqc_remove(df, Pre_QAQC_df, Post_QAQC_df)
}
\arguments{
\item{df}{The input data frame}

\item{Pre_QAQC_df}{PLACEHOLDER}

\item{Post_QAQC_df}{PLACEHOLDER}
}
\value{
qaqc_output_df The final dataframe with NA to indicate bad data
}
\description{
Replaces values that were flagged in the fine QAQC with NA
}
\author{
Kenneth Davidson
}
