% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{sanity_check_course}
\alias{sanity_check_course}
\title{Course QAQC processing.}
\usage{
sanity_check_course(
  df,
  Dates = unique(df$Date),
  Pre_Threshold_Temp = 0.5,
  Pre_Threshold_Num = 20,
  Post_Threshold_Temp = 5,
  Post_Threshold_Num = 10,
  HT_Temp = 0.5
)
}
\arguments{
\item{df}{The input data frame}

\item{Dates}{The dates / date range to process}

\item{Pre_Threshold_Temp}{Difference in temp from mode (during the pre-pulse period) greater than which a point will be considered an outlier}

\item{Pre_Threshold_Num}{Minimum number of outliers during the pre-pulse period before data are flagged as bad}

\item{Post_Threshold_Temp}{Difference in temp from mode (during the post-pulse period) greater than which a point will be considered an outlier}

\item{Post_Threshold_Num}{Minimum number of outliers during the post-pulse period before data are flagged as bad}

\item{HT_Temp}{Temperature threshold for detecting heater pulse}
}
\value{
A list composed of three QAQC data frames
}
\description{
Course QAQC processing. Breaks each pulse interval into three periods, pre-pulse(00:00-02:00), heat (02:01-03:00), and post pre-pulse (02:01-15:00).
Identifies and flags outliers as temperature readings that are greater than given amount above or below the mode temperature reading during that period.
If the number of outliers are exceed a user specified minimum threshold, the whole period is flagged as bad.
Data returned as TRUE indicates an error.
}
\author{
Kenneth Davidson
}
