% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{sanity_check_fine}
\alias{sanity_check_fine}
\title{Fine QAQC processing}
\usage{
sanity_check_fine(
  df,
  Dates = unique(df$Date),
  Pre_Threshold_Temp = 0.5,
  Post_Threshold_Temp = 5
)
}
\arguments{
\item{df}{The input data frame}

\item{Dates}{The dates / date range to process}

\item{Pre_Threshold_Temp}{Difference in temp from mode (during the pre-pulse period) greater than which a point will be considered an outlier}

\item{Post_Threshold_Temp}{Difference in temp from mode (during the post-pulse period) greater than which a point will be considered an outlier}
}
\value{
QAQC list
}
\description{
Fine QAQC processing. Breaks each pulse interval into two periods, pre-pulse(00:00-02:00), and post pre-pulse (02:01-15:00).
Identifies and flags outliers as temperature readings that are greater than given amount above or below the mode temperature reading during that period.
Output is a matrix, equivalent in dimensions to the input data, filled with binary values.
Data returned as TRUE indicates an outlier.
}
\author{
Kenneth Davidson
}
