% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_velocity.R
\name{solve_hrm}
\alias{solve_hrm}
\title{Title}
\usage{
solve_hrm(
  df,
  Dates = unique(df$Date),
  ktherm = rep(0.0025, 5),
  tavg_start = rep(200, 5),
  tavg_end = rep(900, 5),
  sensor_2_dist = rep(0.2, 5),
  sensor_1_dist = rep(-0.2, 5)
)
}
\arguments{
\item{df}{}

\item{Dates}{The dates / date range to process. If omitted all data in df will be processed.}

\item{ktherm}{Wood thermal diffusivity. Supplied as a vector of five values corresponding to Trees 1-5.}

\item{tavg_start}{Beginning of averaging window, equivalent to seconds since start observations. Nominally the heat pulse will occur at/around 180. Supplied as a vector of five values corresponding to Trees 1-5.}

\item{tavg_end}{End of averaging window, equivalent to seconds since start observations. Nominally the observations will run for 900 seconds. Supplied as a vector of five values corresponding to Trees 1-5.}

\item{sensor_2_dist}{Distance in cm for sensor 2 from heater. Negative values are equivalent a position upstream of the heater.}

\item{sensor_1_dist}{Distance in cm for sensor 1 from heater. Negative values are equivalent a position upstream of the heater.}
}
\value{
Data frame with HRM processed data to half hour. Values are in cm hr-1.
}
\description{
Solves the HRM equations for sensors 1 and 2. K can either as a vector or assumed
}
\author{
Kenneth Davidson
}
