% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Spline_Fit}
\alias{Spline_Fit}
\title{A helper function for Heat_Dis_Spline}
\usage{
Spline_Fit(Y_vals, X_vals = Z$TIMESTAMP, X_Vals_New = X)
}
\description{
A helper function for Heat_Dis_Spline
}
\author{
Kenneth Davidson
}
