% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{sanity_check_course}
\alias{sanity_check_course}
\title{Course QAQC processing.}
\usage{
sanity_check_course(
  df,
  Dates = unique(df$Date),
  Pre_Threshold_Temp = 0.5,
  Pre_Threshold_Num = 20,
  Post_Threshold_Temp = 5,
  Post_Threshold_Num = 10,
  HT_Temp = 0.5
)
}
\arguments{
\item{df}{The input data frame}

\item{Dates}{The dates / date range to process}

\item{Pre_Threshold_Temp}{PLACEHOLDER}

\item{Pre_Threshold_Num}{PLACEHOLDER}

\item{Post_Threshold_Temp}{PLACEHOLDER}

\item{Post_Threshold_Num}{PLACEHOLDER}

\item{HT_Temp}{PLACEHOLDER}
}
\value{
QAQC list
}
\description{
Course QAQC processing. Identifies outliers and flags if data
are messy by time period (e.g., 12:00:00-12:01:59) based on number of errant
observations. Also detects if heater fired. Data returned as TRUE indicates
an error. NOTE: This function is VERY SLOW, can be run in parallel to
speed up
}
\author{
Kenneth Davidson
}
