% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{sanity_check_fine}
\alias{sanity_check_fine}
\title{Fine QAQC processing}
\usage{
sanity_check_fine(
  df,
  Dates = unique(df$Date),
  Pre_Threshold_Temp = 0.5,
  Post_Threshold_Temp = 5
)
}
\arguments{
\item{df}{The input data frame}

\item{Dates}{The dates / date range to process}

\item{Pre_Threshold_Temp}{PLACEHOLDER}

\item{Post_Threshold_Temp}{PLACEHOLDER}
}
\value{
QAQC list
}
\description{
Fine QAQC processing. Identifies outliers and flags if data are
messy by individual point. Data returned as TRUE indicates an error.
This function is VERY SLOW, can be run in parallel to speed up
}
\author{
Kenneth Davidson
}
