% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spline_fit}
\alias{spline_fit}
\title{A helper function for Heat_Dis_Spline}
\usage{
spline_fit(Y_vals, X_vals = Z$TIMESTAMP, X_Vals_New = X)
}
\description{
A helper function to fit a spline to a moving window of data.
Used estimate baseline temp of sap post heating
for detection of full pulse duration.
}
\author{
Kenneth Davidson
}
